requirements.txt
imageio==2.4.1
praw
requests
urllib3
keyboard
pillow
beautifulsoup4
instabot
